% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecophysio}
\alias{ecophysio}
\title{Ecophysiological data of seedlings in the Pyrenees}
\format{
A data frame with 191 rows and 14 variables
\describe{
  \item{Stage}{elevation belt at which the plant is located (Montane, Subalpine)}
  \item{Species}{tree species (Abies alba, Betula pendula, Pinus sylvestris, Pinus uncinata)}
  \item{GLI}{light availability measured through Gap Light Index in percentage}
  \item{PB}{Plant biomass in g}
  \item{LB}{Leaf biomass in g}
  \item{SB}{Stem biomass in g}
  \item{RB}{Root biomass in g}
  \item{LWR}{leaf weight ratio (leaf weight/total weight)}
  \item{SWR}{stem weight ratio (stem weight/total weight)}
  \item{RWR}{root weight ratio (root weight/total weight)}
  \item{RMF}{root mass fraction (leaf weight/total weight)}
  \item{RSR}{root:shoot ratio }
  \item{SLA}{specific leaf area in cm2/g(leaf weight/total weight)}
  \item{WatPot}{water potential in bar}

}
}
\source{
<https://annforsci.biomedcentral.com/articles/10.1007/s13595-019-0831-1>
}
\usage{
ecophysio
}
\description{
A dataset containing several ecophysiological and morphological variables for 191 seedlings of 4 tree species
 planted in the Pyrenees. The variables are as follows:
}
\keyword{datasets}
