% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rdpi.R, R/rdpi_matrix.R
\name{rdpi_matrix}
\alias{rdpi_matrix}
\title{rdpi_matrix}
\usage{
rdpi_matrix(data, trait, factor)

rdpi_matrix(data, trait, factor)
}
\arguments{
\item{data}{The dataframe that contains the data}

\item{trait}{The bare (unquoted) name of the column that holds the trait for which to calculate RDPI. Must be numeric}

\item{factor}{the bare (unquoted) name of the column that holds the environmental factor for which we will calculate RDPI. 
By definition, RDPI computes distances between pairs of observations that are at different levels of this factor.}
}
\value{
a vector containing all the calculated rdpi values for each pair of observations that do not belong to the same level of "factor"

a vector containing all the calculated rdpi values for each pair of observations that do not belong to the same level of "factor"
}
\description{
Function to compute a vector containing the relative distance plasticity index values (Valladares et al. 2006) of a given
trait (trait) for a given categorical environmental variable (factor). It calculates rdpi for each pair of observation that does 
not belong to the same level of "factor", and returns a vector containing all the calculated rdpi values.

Function to compute a matrix with the relative distance plasticity index values (Valladares et al. 2006) of a given
trait (trait) for a given categorical environmental variable (factor). It calculates rdpi for each pair of observation that does 
not belong to the same level of "factor", and returns a vector containing all the calculated rdpi values.
}
\examples{
data(ecophysio)
compute_rdpi(ecophysio,SB, Piso)
data(ecophysio)
rdpi_matrix(ecophysio,SB, Piso)
}
